Crate revm_primitives

revm-primitives

EVM primitive types.

Re-exports

pub use kzg::EnvKzgSettings;

pub use alloy_primitives;

pub use alloy_primitives::bytes;

pub use alloy_primitives::hex;

pub use alloy_primitives::hex_literal;

pub use alloy_primitives::ruint;

pub use bitvec;

pub use env::*;

pub use precompile::*;

pub use result::*;

pub use specification::*;

pub use state::*;

pub use utilities::*;

Modules

db

env

hash_map

A hash map implemented with quadratic probing and SIMD lookup.

hash_set

A hash set implemented as a HashMap where the value is ().

kzg

precompile

result

specification

state

utilities

Macros

address

Converts a sequence of string literals containing hex-encoded data into a new Address at compile time.

b256

Converts a sequence of string literals containing hex-encoded data into a new B256 at compile time.

bytes

Converts a sequence of string literals containing hex-encoded data into a new Bytes at compile time.

fixed_bytes

Converts a sequence of string literals containing hex-encoded data into a new FixedBytes at compile time.

hex

Macro for converting sequence of string literals containing hex-encoded data into an array of bytes.

uint

The uint! macro for Uint and Bits literals

Structs

Address

An Ethereum address, 20 bytes in length.

Bytecode

Bytes

Wrapper type around bytes::Bytes to support “0x” prefixed hex strings.

FixedBytes

A byte array of fixed length ([u8; N]).

HashMap

A hash map implemented with quadratic probing and SIMD lookup.

HashSet

A hash set implemented as a HashMap where the value is ().

JumpMap

A map of valid jump destinations.

KzgSettings

Stores the setup and parameters needed for computing KZG proofs.

Log

Enums

BytecodeState

State of the Bytecode analysis.

Constants

BLOB_GASPRICE_UPDATE_FRACTION

Controls the maximum rate of change for blob gas price.

BLOCK_HASH_HISTORY

Number of blocks hashes that EVM can access in the past

GAS_PER_BLOB

Gas consumption of a single data blob (== blob byte size).

MAX_BLOB_GAS_PER_BLOCK

Maximum consumable blob gas for data blobs per block.

MAX_BLOB_NUMBER_PER_BLOCK

Max number of blobs per block

MAX_CODE_SIZE

EIP-170: Contract code size limit By default limit is 0x6000 (~25kb)

MAX_INITCODE_SIZE

EIP-3860: Limit and meter initcode

MIN_BLOB_GASPRICE

Minimum gas price for data blobs.

PRECOMPILE3

Precompile 3 is special in few places

TARGET_BLOB_GAS_PER_BLOCK

Target consumable blob gas for data blobs per block (for 1559-like pricing).

TARGET_BLOB_NUMBER_PER_BLOCK

Target number of the blob per block.

VERSIONED_HASH_VERSION_KZG

First version of the blob.

Type Aliases

B256

32-byte fixed byte-array type.

I256

256-bit signed integer type, consisting of 4, 64-bit limbs.

U256

Uint for 256 bits.