Struct revm_primitives::bits::B160

#[repr(C)]
pub struct B160(pub [u8; 20]);

revm 160 bits type.

Tuple Fields
0: [u8; 20]
Implementations

impl B160

pub const fn repeat_byte(byte: u8) -> B160

Returns a new fixed hash where all bits are set to the given byte.

pub const fn zero() -> B160

Returns a new zero-initialized fixed hash.

pub const fn len_bytes() -> usize

Returns the size of this hash in bytes.

pub fn as_bytes(&self) -> &[u8]

Extracts a byte slice containing the entire fixed hash.

pub fn as_bytes_mut(&mut self) -> &mut [u8]

Extracts a mutable byte slice containing the entire fixed hash.

pub const fn as_fixed_bytes(&self) -> &[u8; 20]

Extracts a reference to the byte array containing the entire fixed hash.

pub fn as_fixed_bytes_mut(&mut self) -> &mut [u8; 20]

Extracts a reference to the byte array containing the entire fixed hash.

pub const fn to_fixed_bytes(self) -> [u8; 20]

Returns the inner bytes array.

pub fn as_ptr(&self) -> *const u8

Returns a constant raw pointer to the value.

pub fn as_mut_ptr(&mut self) -> *mut u8

Returns a mutable raw pointer to the value.

pub fn assign_from_slice(&mut self, src: &[u8])

Assign the bytes from the byte slice src to self.

Note

The given bytes are interpreted in big endian order.

Panics

If the length of src and the number of bytes in self do not match.

pub fn from_slice(src: &[u8]) -> Self

Create a new fixed-hash from the given slice src.

Note

The given bytes are interpreted in big endian order.

Panics

If the length of src and the number of bytes in Self do not match.

pub fn covers(&self, b: &Self) -> bool

Returns true if all bits set in b are also set in self.

pub fn is_zero(&self) -> bool

Returns true if no bits are set.

impl B160

Utilities using the byteorder crate.

pub fn to_low_u64_be(&self) -> u64

Returns the lowest 8 bytes interpreted as big-endian.

Note

For hash type with less than 8 bytes the missing bytes are interpreted as being zero.

pub fn to_low_u64_le(&self) -> u64

Returns the lowest 8 bytes interpreted as little-endian.

Note

For hash type with less than 8 bytes the missing bytes are interpreted as being zero.

pub fn to_low_u64_ne(&self) -> u64

Returns the lowest 8 bytes interpreted as native-endian.

Note

For hash type with less than 8 bytes the missing bytes are interpreted as being zero.

pub fn from_low_u64_be(val: u64) -> Self

Creates a new hash type from the given u64 value.

Note

The given u64 value is interpreted as big endian.

Ignores the most significant bits of the given value if the hash type has less than 8 bytes.

pub fn from_low_u64_le(val: u64) -> Self

Creates a new hash type from the given u64 value.

Note

The given u64 value is interpreted as little endian.

Ignores the most significant bits of the given value if the hash type has less than 8 bytes.

pub fn from_low_u64_ne(val: u64) -> Self

Creates a new hash type from the given u64 value.

Note

The given u64 value is interpreted as native endian.

Ignores the most significant bits of the given value if the hash type has less than 8 bytes.

impl B160

Utilities using the rand crate.

pub fn randomize_using<R>(&mut self, rng: &mut R)where R: Rng + ?Sized,

Assign self to a cryptographically random value using the given random number generator.

pub fn randomize(&mut self)

Assign self to a cryptographically random value.

pub fn random_using<R>(rng: &mut R) -> Selfwhere R: Rng + ?Sized,

Create a new hash with cryptographically random content using the given random number generator.

pub fn random() -> Self

Create a new hash with cryptographically random content.