revm_primitives
Module eip4844

EIP-4844 constants

Constants

BLOB_BASE_FEE_UPDATE_FRACTION_CANCUN

Controls the maximum rate of change for blob gas price

BLOB_BASE_FEE_UPDATE_FRACTION_PRAGUE

Controls the maximum rate of change for blob gas price

GAS_PER_BLOB

Gas consumption of a single data blob (== blob byte size)

MAX_BLOB_GAS_PER_BLOCK_CANCUN

Maximum consumable blob gas for data blobs per block

MAX_BLOB_GAS_PER_BLOCK_PRAGUE

Maximum consumable blob gas for data blobs per block

MAX_BLOB_NUMBER_PER_BLOCK_CANCUN

Max number of blobs per block

MAX_BLOB_NUMBER_PER_BLOCK_PRAGUE

Max number of blobs per block

MIN_BLOB_GASPRICE

Min blob gas price

TARGET_BLOB_GAS_PER_BLOCK_CANCUN

Target consumable blob gas for data blobs per block (for 1559-like pricing)

TARGET_BLOB_GAS_PER_BLOCK_PRAGUE

Target consumable blob gas for data blobs per block (for 1559-like pricing)

TARGET_BLOB_NUMBER_PER_BLOCK_CANCUN

Target number of the blob per block

TARGET_BLOB_NUMBER_PER_BLOCK_PRAGUE

Target number of the blob per block

VERSIONED_HASH_VERSION_KZG

First version of the blob