Module revm_primitives::eip7702

Re-exports

pub use authorization_list::AuthorizationList;

pub use bytecode::Eip7702Bytecode;

pub use bytecode::Eip7702DecodeError;

pub use bytecode::EIP7702_MAGIC;

pub use bytecode::EIP7702_MAGIC_BYTES;

pub use bytecode::EIP7702_MAGIC_HASH;

pub use bytecode::EIP7702_VERSION;

Modules

authorization_list

bytecode

Structs

Authorization

An unsigned EIP-7702 authorization.

PrimitiveSignature

An Ethereum ECDSA signature.

RecoveredAuthorization

A recovered authorization.

SignedAuthorization

A signed EIP-7702 authorization.

Enums

RecoveredAuthority

Represents the outcome of an attempt to recover the authority from an authorization. It can either be valid (containing an Address) or invalid (indicating recovery failure).

Constants

PER_AUTH_BASE_COST

PER_EMPTY_ACCOUNT_COST

Cost of creating authorized account that was previously empty.