Crate revm_primitives

revm-primitives

EVM primitive types.

Re-exports

pub use alloy_primitives;

pub use alloy_primitives::bytes;

pub use alloy_primitives::hex;

pub use alloy_primitives::ruint;

pub use constants::*;

Modules

constants

Global constants for the EVM

eip170

EIP-170: Contract code size limit

eip4844

EIP-4844 constants

eip7702

EIP-7702 constants

eip7823

EIP-7823 Set upper bounds for MODEXP

eip7825

EIP-7825: Transaction Gas Limit Cap
Introduce a protocol-level cap on the maximum gas used by a transaction to 30 million.

eip7918

Constants for EIP-7918: Blob base fee bounded by execution cost

eof

EOF related constants and functions.

hardfork

hash_map

A hash map implemented with quadratic probing and SIMD lookup.

hash_set

A hash set implemented as a HashMap where the value is ().

map

Re-exports of map types and utilities.

Macros

address

Converts a sequence of string literals containing hex-encoded data into a new Address at compile time.

b256

Converts a sequence of string literals containing hex-encoded data into a new B256 at compile time.

bytes

Converts a sequence of string literals containing hex-encoded data into a new Bytes at compile time.

fixed_bytes

Converts a sequence of string literals containing hex-encoded data into a new FixedBytes at compile time.

hex

Macro for converting sequence of string literals containing hex-encoded data into an array of bytes.

Structs

Address

An Ethereum address, 20 bytes in length.

Bytes

Wrapper type around bytes::Bytes to support “0x” prefixed hex strings.

FixedBytes

A byte array of fixed length ([u8; N]).

Log

A log consists of an address, and some log data.

LogData

An Ethereum event log object.

Enums

TxKind

The to field of a transaction. Either a target address, or empty for a contract creation.

Functions

keccak256

Simple interface to the Keccak-256 hash function.

Type Aliases

B256

32-byte fixed byte-array type.

HashMap

A HashMap using the default hasher.

HashSet

A HashSet using the default hasher.

I128

128-bit signed integer type, consisting of 2, 64-bit limbs.

I256

256-bit signed integer type, consisting of 4, 64-bit limbs.

StorageKey

type alias for storage keys

StorageValue

type alias for storage values

U128

128-bit unsigned integer type, consisting of 2, 64-bit limbs.

U256

256-bit unsigned integer type, consisting of 4, 64-bit limbs.