Crate revm_primitives

Re-exports

pub use bits::B160;

pub use bits::B256;

pub use bytes;

pub use hex;

pub use hex_literal;

pub use bitvec;

pub use log::Log;

pub use ruint;

pub use bytecode::*;

pub use env::*;

pub use precompile::*;

pub use result::*;

pub use specification::*;

pub use state::*;

pub use utilities::*;

Modules

bits

bytecode

db

env

hash_map

A hash map implemented with quadratic probing and SIMD lookup.

log

precompile

result

specification

state

utilities

Macros

uint

The uint! macro for Uint literals

Structs

Bytes

A cheaply cloneable and sliceable chunk of contiguous memory.

HashMap

A hash map implemented with quadratic probing and SIMD lookup.

Type Definitions

Address

Address type is last 20 bytes of hash of ethereum account

Hash

Hash, in Ethereum usually kecack256.

U256

Uint for 256 bits.